refactor(src) add src module, remove all
