+++ src/api/routers/cameras.py
