+++ src/database/repositories/cameras.py
