+++ src/models/files.py
@@ -1,5 +1,5 @@
 from sqlalchemy import func
-from .base import Base
+from ..database.engine.base import Base
 from sqlalchemy.orm import Mapped, mapped_column
 from datetime import datetime
