+++ src/models/violations.py
@@ -1,5 +1,5 @@
 from sqlalchemy import UniqueConstraint
-from .base import Base
+from ..database.engine.base import Base
 from sqlalchemy.orm import Mapped, mapped_column
 
 class ViolationORM(Base):
