+++ src/schemas/cameras.py
